Re-factor Challenge Display page.
Formatting the styling for each challenges in a much bearable looking.

Tweaking around the PHP functionaily a bit to compatible with the new style changes.

Removed specific 'Google Keyword Search' page as it has been migrated to this page.

Added  proper HTML5 format.

Correct spacing implemented for a better way to read the code blocks.