Swapped from using LCD display to MiniTFT display.
Temporary swap as the LCD is not working properly.

Still prefer to use the LCD more but that can be done later on.

Added some missing libraries from last commit.

Although the display text is still not working (probably becasue the device itself has some problems with the screen), I got the input respone when clicking on each buttons and joysticks from the MiniTFT.