backupDieselGenerators Challenge Complete
The challenge is complete and funtional. The PHP team needs to create a webpage for it and allow users to download the files from /Challenge_Files/versions/

Currently, it lacks security and the webservers password is held in cleartext within the code.

There is also an old proof of concept SSH connection script in there. Hasnt been compiled.