Added 'test.cpp' file.
This file is to checking if the TFT display working or not. Currently, all of the functions in there is worked (Debugged text also printed successfully in Serial Monitor) but the actual module is not reflecting it. Which means that it is a possibility that the module itself has issues on it a.k.a hardware issues.

Fix warning in 'display.cpp' file where it was forced to sent a 'char' text type while the actual text is a 'const char' type.