Working on fetching data from module to server and vice versa.
I got it worked as outputting the data as JSON in terminal.

Not sure why it is not sending the data the the website when I entered the correct flag. Maybe it have something to do with the server database (I have no access over it so...).

Wrapped the Wi-fi connection function into a custom call function, just like the text display.

Add some time delays to prevent memory got filled up quickly as requests sending repaidly.

Include library for JSON data as well.