refactor(module-register-page): adapt better UI / UX form design for 'Module Register' page
We have received feedbacks from our clients about their user experience when navigating through this page is very confusing as well as cluttered, etc. Because of so, Our UI / UX design team have had a quick meeting on what can be the best solution to solve this. As a result, this is the new design that out team has concluded after the meeting and implemeting. For a quick summary:

- Instead of each form will have its own line (which considered confusing for our clients), we have adjust them into a grid-like system that could provides a better UI for our clients.
- Also, we add comments to explains what kinds of forms we are referring to for each section in this grid-like system so that it will be much easier for our developers to upgrade / adjust anything in the future.
- Adjust the title name into a more accurate naming that reflects the purpose of this page.