Fix navbar issue
Due to the logo image being put inside an <a> tag to embed a link to home within it, the element for the image was taking up all free space within the navbar, stopping items from aligning to the left. Removing the <a> tag and returning it to how it was in the original has fixed the alignment issue.
Some navbar items are not in line, however it is all functional