Added DatabaseToMQTT.py
needs to be made into a daemon or some kinda persistant background process on the server.

It reads updates from the database via the binlog, so it doesnt have to constantly query it. Any changes to a row in the Registered Modules table is detected.

The 'Module' column of that row is used as the MQTT publish topic, and the message is the 'Current Output' column